+++ common/src/main/java/com/github/skillfi/tensura_mf/recipe/input/MagicIncubationRecipeInput.java
@@ -3,7 +3,7 @@ package com.github.skillfi.tensura_mf.recipe.input;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.RecipeInput;
 
-public record MagicIncubationRecipeInput(ItemStack item, int magicAmount) implements RecipeInput {
+public record MagicIncubationRecipeInput(ItemStack item, Float magicAmount) implements RecipeInput {
     @Override
     public ItemStack getItem(int i) {
         return this.item;
